move SQLTokenGenerateEngine to builder package
